requirements.txt
# run pip install -r requirements.txt
pyelftools>=0.29
capstone>=4.0.2